+++ packages/cryptography/src/index.js
@@ -3,5 +3,6 @@ export { default as KeyList } from "./KeyList.js";
 export { default as PrivateKey } from "./PrivateKey.js";
 export { default as PublicKey } from "./PublicKey.js";
 export { default as Mnemonic } from "./Mnemonic.js";
+export { default as BadKeyError } from "./BadKeyError.js";
 export { default as BadMnemonicError } from "./BadMnemonicError.js";
 export { default as BadMnemonicReason } from "./BadMnemonicReason.js";